[#1596] run formatting over changes
