+++ src/openzaak/components/documenten/tests/import/test_import_documents.py
@@ -1,26 +1,38 @@
 import csv
 from io import StringIO
 from pathlib import Path
-from uuid import uuid4
 from unittest.mock import patch
+from uuid import uuid4
 
 from django.db import IntegrityError, OperationalError
 from django.test import TestCase, override_settings
 
-from openzaak.components.catalogi.tests.factories.informatie_objecten import InformatieObjectTypeFactory
+from openzaak.components.catalogi.tests.factories.informatie_objecten import (
+    InformatieObjectTypeFactory,
+)
+from openzaak.components.documenten.import_utils import DocumentRow
 from openzaak.components.documenten.models import EnkelvoudigInformatieObject
-from openzaak.components.documenten.tasks import DocumentRow, import_documents
-from openzaak.components.documenten.tests.factories import DocumentRowFactory, EnkelvoudigInformatieObjectFactory
+from openzaak.components.documenten.tasks import import_documents
+from openzaak.components.documenten.tests.factories import (
+    DocumentRowFactory,
+    EnkelvoudigInformatieObjectFactory,
+)
 from openzaak.components.zaken.tests.factories import ZaakFactory
-from openzaak.import_data.models import ImportRowResultChoices, ImportStatusChoices, ImportTypeChoices
-from openzaak.import_data.tests.factories import ImportFactory, get_informatieobjecttype_url
+from openzaak.import_data.models import (
+    ImportRowResultChoices,
+    ImportStatusChoices,
+    ImportTypeChoices,
+)
+from openzaak.import_data.tests.factories import (
+    ImportFactory,
+    get_informatieobjecttype_url,
+)
 from openzaak.import_data.tests.utils import ImportTestFileMixin
 from openzaak.utils.fields import get_default_path
 
 
 @override_settings(
-    IMPORT_DOCUMENTEN_BASE_DIR="/tmp/import/",
-    IMPORT_DOCUMENTEN_BATCH_SIZE=2
+    IMPORT_DOCUMENTEN_BASE_DIR="/tmp/import/", IMPORT_DOCUMENTEN_BATCH_SIZE=2
 )
 class ImportDocumentTestCase(ImportTestFileMixin, TestCase):
     def test_simple_import(self):
@@ -50,7 +62,7 @@ class ImportDocumentTestCase(ImportTestFileMixin, TestCase):
             4: DocumentRowFactory(
                 informatieobjecttype=informatieobjecttype_url,
                 bestandspad="test-file-4.docx",
-                zaak_id=str(zaken[2].uuid)
+                zaak_id=str(zaken[2].uuid),
             ),
         }
 
@@ -94,9 +106,7 @@ class ImportDocumentTestCase(ImportTestFileMixin, TestCase):
         self.assertEqual(DocumentRow.export_headers, rows[0])
 
         self.assertTrue(
-            all(
-                (row[-1] == ImportRowResultChoices.imported.label) for row in rows[1:]
-            )
+            all((row[-1] == ImportRowResultChoices.imported.label) for row in rows[1:])
         )
 
         # no comments on all the rows
@@ -113,7 +123,7 @@ class ImportDocumentTestCase(ImportTestFileMixin, TestCase):
             1: DocumentRowFactory(
                 informatieobjecttype=informatieobjecttype_url,
                 bestandspad="test-file-1.docx",
-                uuid="foobar" # Note the incorrect uuid
+                uuid="foobar",  # Note the incorrect uuid
             ),
             2: DocumentRowFactory(
                 informatieobjecttype=informatieobjecttype_url,
@@ -168,7 +178,7 @@ class ImportDocumentTestCase(ImportTestFileMixin, TestCase):
         self.assertEqual(len(rows), 5)
         self.assertEqual(DocumentRow.export_headers, rows[0])
 
-        success_rows = (3, 4, 5) # note the header row
+        success_rows = (3, 4, 5)  # note the header row
 
         for row_index, row in enumerate(rows, start=1):
             if row_index == 1:
@@ -176,24 +186,20 @@ class ImportDocumentTestCase(ImportTestFileMixin, TestCase):
 
             with self.subTest(row_index=row_index, row=row):
                 if row_index not in success_rows:
-                    self.assertEqual(
-                        row[-1], ImportRowResultChoices.not_imported.label
-                    )
+                    self.assertEqual(row[-1], ImportRowResultChoices.not_imported.label)
 
                     self.assertIn("not a valid UUID", row[-2])
 
                     continue
 
-                self.assertEqual(
-                    row[-1], ImportRowResultChoices.imported.label
-                )
+                self.assertEqual(row[-1], ImportRowResultChoices.imported.label)
 
                 self.assertEqual(row[-2], "")
 
     @patch("openzaak.components.documenten.tasks.uuid4")
     @patch(
         "openzaak.components.documenten.tasks.EnkelvoudigInformatieObject.objects.bulk_create",
-        autospec=True
+        autospec=True,
     )
     def test_database_connection_loss(self, mocked_bulk_create, mocked_uuid):
         informatieobjecttype = InformatieObjectTypeFactory(
@@ -253,10 +259,7 @@ class ImportDocumentTestCase(ImportTestFileMixin, TestCase):
             size=2, informatieobjecttype=informatieobjecttype
         )
 
-        mocked_bulk_create.side_effect = (
-            random_eios,
-            OperationalError
-        )
+        mocked_bulk_create.side_effect = (random_eios, OperationalError)
 
         mocked_uuid.side_effect = [eio.uuid for eio in random_eios]
 
@@ -282,7 +285,7 @@ class ImportDocumentTestCase(ImportTestFileMixin, TestCase):
         error_rows = (4, 5)
 
         for row_index, row in enumerate(rows, start=1):
-            if row_index == 1: # header row
+            if row_index == 1:  # header row
                 continue
 
             with self.subTest(row_index=row_index, row=row):
@@ -306,7 +309,7 @@ class ImportDocumentTestCase(ImportTestFileMixin, TestCase):
     @patch("openzaak.components.documenten.tasks.uuid4")
     @patch(
         "openzaak.components.documenten.tasks.EnkelvoudigInformatieObject.objects.bulk_create",
-        autospec=True
+        autospec=True,
     )
     def test_integrity_error(self, mocked_bulk_create, mocked_uuid):
         informatieobjecttype = InformatieObjectTypeFactory(
@@ -364,10 +367,7 @@ class ImportDocumentTestCase(ImportTestFileMixin, TestCase):
             size=2, informatieobjecttype=informatieobjecttype
         )
 
-        mocked_bulk_create.side_effect = (
-            IntegrityError,
-            random_eios
-        )
+        mocked_bulk_create.side_effect = (IntegrityError, random_eios)
 
         mocked_uuid.side_effect = [eio.uuid for eio in random_eios]
 
@@ -393,7 +393,7 @@ class ImportDocumentTestCase(ImportTestFileMixin, TestCase):
         error_rows = (2, 3)
 
         for row_index, row in enumerate(rows, start=1):
-            if row_index == 1: # header row
+            if row_index == 1:  # header row
                 continue
 
             with self.subTest(row_index=row_index, row=row):
@@ -421,7 +421,7 @@ class ImportDocumentTestCase(ImportTestFileMixin, TestCase):
             1: DocumentRowFactory(
                 informatieobjecttype=informatieobjecttype_url,
                 bestandspad="test-file-1.docx",
-                zaak_id="36df518e-7dff-4af3-be96-ccca0c6e6a2e", # unknown
+                zaak_id="36df518e-7dff-4af3-be96-ccca0c6e6a2e",  # unknown
             ),
             2: DocumentRowFactory(
                 informatieobjecttype=informatieobjecttype_url,
@@ -435,7 +435,7 @@ class ImportDocumentTestCase(ImportTestFileMixin, TestCase):
             4: DocumentRowFactory(
                 informatieobjecttype=informatieobjecttype_url,
                 bestandspad="test-file-4.docx",
-                zaak_id=str(zaken[1].uuid)
+                zaak_id=str(zaken[1].uuid),
             ),
         }
 
@@ -481,7 +481,7 @@ class ImportDocumentTestCase(ImportTestFileMixin, TestCase):
         error_row = 2
 
         for row_index, row in enumerate(rows, start=1):
-            if row_index == 1: # header row
+            if row_index == 1:  # header row
                 continue
 
             with self.subTest(row_index=row_index, row=row):
@@ -571,7 +571,7 @@ class ImportDocumentTestCase(ImportTestFileMixin, TestCase):
         error_rows = (2, 3)
 
         for row_index, row in enumerate(rows, start=1):
-            if row_index == 1: # header row
+            if row_index == 1:  # header row
                 continue
 
             with self.subTest(row_index=row_index, row=row):
@@ -579,9 +579,14 @@ class ImportDocumentTestCase(ImportTestFileMixin, TestCase):
                     self.assertEqual(row[-1], ImportRowResultChoices.not_imported.label)
 
                     if row_index == 2:
-                        self.assertIn("Unable to couple row 2 to ZAAK b0f3681d-945a-4b30-afcb-12cad0a3eeaf:", row[-2])
+                        self.assertIn(
+                            "Unable to couple row 2 to ZAAK b0f3681d-945a-4b30-afcb-12cad0a3eeaf:",
+                            row[-2],
+                        )
                     else:
-                        self.assertIn("Unable to load row due to database error", row[-2])
+                        self.assertIn(
+                            "Unable to load row due to database error", row[-2]
+                        )
 
                     continue
 
@@ -663,14 +668,17 @@ class ImportDocumentTestCase(ImportTestFileMixin, TestCase):
         self.assertEqual(DocumentRow.export_headers, rows[0])
 
         for row_index, row in enumerate(rows, start=1):
-            if row_index == 1: # header row
+            if row_index == 1:  # header row
                 continue
 
             with self.subTest(row_index=row_index, row=row):
                 self.assertEqual(row[-1], ImportRowResultChoices.not_imported.label)
 
                 if row_index == 2:
-                    self.assertIn("Unable to couple row 2 to ZAAK b0f3681d-945a-4b30-afcb-12cad0a3eeaf:", row[-2])
+                    self.assertIn(
+                        "Unable to couple row 2 to ZAAK b0f3681d-945a-4b30-afcb-12cad0a3eeaf:",
+                        row[-2],
+                    )
                 else:
                     self.assertIn("Unable to load row due to database error", row[-2])
