+++ src/openzaak/import_data/tests/factories.py
@@ -39,8 +39,7 @@ def get_informatieobjecttype_url(
     instance_url = reverse(
         "informatieobjecttype-detail",
         kwargs=dict(
-            uuid=instance.uuid,
-            version=settings.REST_FRAMEWORK["DEFAULT_VERSION"],
+            uuid=instance.uuid, version=settings.REST_FRAMEWORK["DEFAULT_VERSION"],
         ),
     )
